Fix resource cleanup when zone_id config set
If zone_id set in config the recordset scenario checks fail
cleanup as it tried to delete a tld that is not setup.

+++ designate_tempest_plugin/tests/scenario/v2/test_recordsets.py
@@ -75,7 +75,9 @@ class RecordsetsTest(base.BaseDnsV2Test):
 
     @classmethod
     def resource_cleanup(cls):
-        cls.admin_tld_client.delete_tld(cls.class_tld[1]['id'])
+        zone_id = CONF.dns.zone_id
+        if not zone_id:
+            cls.admin_tld_client.delete_tld(cls.class_tld[1]['id'])
         super(RecordsetsTest, cls).resource_cleanup()
 
     def _test_create_and_delete_records_on_existing_zone(
